artifacts/releases/devlaberp-release-0.0.2/devlaberp-release-0.0.2/requirements.txt
Django==4.2.22
dj-database-url==2.3.0
python-decouple==3.8
Pillow==11.1.0
whitenoise==6.9.0
waitress>=3.0,<4
psycopg[binary]>=3.1,<4
